+++ src/main/java/com/laserfiche/api/client/httphandlers/OAuthClientCredentialsHandler.java
@@ -12,6 +12,7 @@ public class OAuthClientCredentialsHandler implements HttpRequestHandler {
     private String accessToken;
     private final String servicePrincipalKey;
     private final AccessKey accessKey;
+    private final String scope;
     private final TokenClient client;
 
     /**
@@ -20,8 +21,19 @@ public class OAuthClientCredentialsHandler implements HttpRequestHandler {
      * @param accessKey The access key exported from the Laserfiche Developer Console.
      */
     public OAuthClientCredentialsHandler(String servicePrincipalKey, AccessKey accessKey) {
+        this(servicePrincipalKey, accessKey, null);
+    }
+
+    /**
+     * Creates a new Laserfiche Cloud OAuth client credentials HTTP handler.
+     * @param servicePrincipalKey The service principal key created for the service principal from the Laserfiche Account Administration.
+     * @param accessKey The access key exported from the Laserfiche Developer Console.
+     * @param scope The requested space-delimited scopes for the access token.
+     */
+    public OAuthClientCredentialsHandler(String servicePrincipalKey, AccessKey accessKey, String scope) {
         this.servicePrincipalKey = servicePrincipalKey;
         this.accessKey = accessKey;
+        this.scope = scope;
         client = new TokenClientImpl(accessKey.getDomain());
     }
 
@@ -29,7 +41,7 @@ public class OAuthClientCredentialsHandler implements HttpRequestHandler {
     public BeforeSendResult beforeSend(com.laserfiche.api.client.httphandlers.Request request) {
         BeforeSendResult result = new BeforeSendResult();
         if (accessToken == null || accessToken.equals("")) {
-            GetAccessTokenResponse tokenResponse  = client.getAccessTokenFromServicePrincipal(servicePrincipalKey, accessKey);
+            GetAccessTokenResponse tokenResponse  = client.getAccessTokenFromServicePrincipal(servicePrincipalKey, accessKey, scope);
             accessToken = tokenResponse.getAccessToken();
         }
         request.headers().append("Authorization", "Bearer " + accessToken);